Cập nhập cấu trúc project
-  Cập nhập tính năng đọc file
