+++ Quan-Ly-Vat-Tu/BienDoHoa.h
@@ -1,0 +1,35 @@
+// FILE CHUA CAC BIEN DO HOA ( TOA DO ) -> DAT BIEN TOAN CUC 
+#pragma once
+const int ENTER = 13;  
+const int ESC = 27;  
+const int BACK_SPACE = 8; 
+const int SPACE = 32; 
+
+const int DODAIMANHINH = 1536; 
+const int DORONGMANHINH = 864; 
+const int  BACKGROUP = 7; 
+const int COLOR_MENU_CAP1 = 2; 
+const int COLOR_HIGHT_LIGHT = 8;     
+const int COLOR_TEXT = 10; 
+const int COLOR_BODER_INPUT = 2;  
+
+// cap do menu 1
+const int ID_VT = 0 ;
+const int ID_NV = 1;
+const int ID_HD = 2;
+const int ID_DT = 3;
+
+// ID cua cac menu con cua ID_VT 
+const int ID_THEM_VT= 4;
+const int ID_HIEUCHINH_VT= 5;
+const int ID_IN_VT = 6 ;
+
+// mau cho cac nut con 
+const int COLOR_BTN_SUB  = 6;  // mau cua cac nut con 
+const int COLOR_TEXT_BTN_SUB = 10;
+const int COLOR_HIGHT_BTN_SUB = 8;
+
+// dinh nghia mau nen o Nhap 
+const int COLOR_BACKGROUP_INPUT = 7;
+const int CACH_LE_O_NHAP = 10;   // cach cai thanh o nhap ra bao nhieu px
+const int COLOR_TEXT_INPUT = 2;
