+++ Quan-Ly-Vat-Tu/MenuQLVT.h
@@ -1,0 +1,81 @@
+#pragma once
+using namespace std; 
+#include "BstVatTu.h"
+#include "ListCTHD.h"
+#include "ListHoaDon.h"
+#include "dsNhanVien.h"
+#include "CoSoDoHoa.h"
+#include "DoHoaVT.h"
+
+
+
+
+// ===================================MENU ROOT=========================================================
+void menuCha(BstVatTu &dsVatTu) {  
+	int index = -1;  
+	int doDai = DODAIMANHINH / 4 - 16; 
+	const int chieuCaoMenuCap1 = 40;  
+	NutBam vatTu(0,0, doDai,chieuCaoMenuCap1,COLOR_MENU_CAP1,COLOR_HIGHT_LIGHT,COLOR_TEXT,"VAT TU");
+	NutBam nhanVien(doDai, 0, doDai, chieuCaoMenuCap1,COLOR_MENU_CAP1,COLOR_HIGHT_LIGHT,COLOR_TEXT, "NHAN VIEN");
+	NutBam hoaDon( doDai*2 , 0, doDai , chieuCaoMenuCap1,COLOR_MENU_CAP1,COLOR_HIGHT_LIGHT,COLOR_TEXT, "HOA DON");
+	NutBam doanhThu( doDai * 3, 0, doDai, chieuCaoMenuCap1,COLOR_MENU_CAP1,COLOR_HIGHT_LIGHT,COLOR_TEXT, "DOANH THU");
+	vatTu.veNut(); 
+	nhanVien.veNut(); 
+	hoaDon.veNut(); 
+	doanhThu.veNut(); 
+	bool kiemTra[4] = { false }; 
+	int mx = -1;  
+	int my = -1;
+	while (true) {
+		if (index != -1) {
+			if (index == 0) {
+				index = -1 ; 
+				hienThiTinhNangVatTu(dsVatTu,index,nhanVien,hoaDon,doanhThu); 
+			}
+			else if (index == 1) {
+			
+			}
+		}
+		if (ismouseclick(WM_MOUSEMOVE)) {
+			getmouseclick(WM_MOUSEMOVE, mx, my);  
+			vatTu.kiemTraVeLai(mx,my,kiemTra[0]);   
+			nhanVien.kiemTraVeLai(mx,my,kiemTra[1]); 
+			hoaDon.kiemTraVeLai(mx,my,kiemTra[2]); 
+			doanhThu.kiemTraVeLai(mx,my,kiemTra[3]); 
+		}
+		else if (ismouseclick(WM_LBUTTONDOWN)) {  
+			int xclick = -1;
+			int yclick = -1; 
+			getmouseclick(WM_LBUTTONDOWN, xclick, yclick);  
+			if (vatTu.isMouseHover(xclick, yclick)) {
+				vatTu.duocTroVao = true; 
+				vatTu.veNut(); 
+				hienThiTinhNangVatTu(dsVatTu,index,nhanVien,hoaDon,doanhThu);  
+			}
+			else if (hoaDon.isMouseHover(xclick, yclick)) {
+				hoaDon.duocTroVao = true;
+				hoaDon.veNut(); 
+			}
+			else if (nhanVien.isMouseHover(xclick, yclick)) {
+				nhanVien.duocTroVao = true;
+				nhanVien.veNut();
+			}
+			else if (doanhThu.isMouseHover(xclick,yclick)) {
+				doanhThu.duocTroVao = true;
+				doanhThu.veNut();
+			}
+		}
+		 delay(1);  
+	}	
+}
+
+// =================================MENU NHAN THAM SO==============================================
+void menuQuanLyVatTu(BstVatTu &dsVatTu) {  
+	setfillstyle(SOLID_FILL, BACKGROUP);  
+	bar(0, 0, 3000, 3000);
+	menuCha(dsVatTu); 
+}
+
+
+
+
