+++ Quan-Ly-Vat-Tu/data/dataCTHD.txt
@@ -1,0 +1,4 @@
+3
+VT01,2,2,2
+VT02,3,3,3
+VT04,3,3,5
