--- Quan-Ly-Vat-Tu/data/duLieuChuanVatTu.txt
@@ -1,11 +1,0 @@
-10
-VT05,Vat tu 0,kg,12
-VT02,Vat tu 0,kg,12
-VT09,Vat tu 0,kg,12
-VT01,Vat tu 0,kg,12
-VT08,Vat tu 0,kg,12
-VT07,Vat tu 0,kg,12
-VT11,Vat tu 0,kg,12
-VT33,Vat tu 0,kg,12
-VT98,Vat tu 0,kg,12
-VT22,Vat tu 0,kg,12
