+++ Quan-Ly-Vat-Tu/dsNhanVien.h
@@ -1,12 +1,11 @@
 #pragma once
-#include "autoLoad.h"
 using namespace std;
-
-#define NHANVIENTOIDA 500 // n //sua
+#include "Tienich.h"
+#define NHANVIENTOIDA 500 
 
 
 struct NhanVien {
-    string maNV;  // maNV
+    string maNV;  
     string ho;
     string ten;
     int phai;   // 0. nam 1. nu 
@@ -114,8 +113,8 @@ struct NhanVien {
 
 class dsNhanVien {
 private:
-	int soLuong;  // so luong thuc te. 
-    NhanVien *nv[NHANVIENTOIDA];  // 
+	int soLuong;  
+    NhanVien *nv[NHANVIENTOIDA]; 
 public:
 	dsNhanVien()
 	{
